requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication
python-keycloak==3.9.0
PyJWT==2.8.0
python-jose[cryptography]==3.3.0

# AI/ML Services
openai==1.10.0
langchain==0.1.4
langchain-community==0.0.16
langchain-openai==0.0.5
faiss-cpu==1.7.4
sentence-transformers==2.3.1

# STT/TTS
openai-whisper==20231117
elevenlabs==0.2.26
soundfile==0.12.1
pydub==0.25.1

# Document Processing
pypdf==3.17.4
python-docx==1.1.0
pandas==2.1.4
openpyxl==3.1.2

# gRPC
grpcio==1.60.0
grpcio-tools==1.60.0

# WebRTC
aiortc==1.6.0

# Monitoring
prometheus-client==0.19.0
boto3==1.34.30

# Optimization
bitsandbytes==0.42.0
torch==2.1.2

# Utilities
httpx==0.26.0
python-dotenv==1.0.0
redis==5.0.1
